Refactor submitSearchTerm function
